+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Altered watcher ID generation to avoid potential collisions
 - Added optional $onStart callback parameter to Reactor::run() implementations
+- Added Scala-like Future\Promise implementation
+
+> **BC BREAKS**: *none*
 
 v0.3.0
 ------